% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_lists.R
\name{data_to_lists}
\alias{data_to_lists}
\title{Data frame to lists}
\usage{
data_to_lists(data, class_var = "class", y_var = "y")
}
\arguments{
\item{data}{data frame with (at least) all the variables in the formula as 
well as the class identifier variable}

\item{class_var}{the name of the class variable in the data frame}

\item{y_var}{the name of the dependent variable in the data frame}
}
\value{
A list with the following components
\item{Y}{list of size \eqn{K}. Its kth entry is a vector of size \eqn{n_k}}
\item{X}{list of size \eqn{K}. Its kth entry is a matrix of size \eqn{n_{k} \times p}}
}
\description{
Takes a data frame and return the lists required for \code{muliclass_reg()} to work
}
\examples{
p = 7
k = 2
n = 20
beta = c(1,2)
set.seed(1)
X = list(matrix(rnorm(p*n), ncol = p), matrix(rnorm(p*n), ncol = p))
Y = list(rnorm(n), rnorm(n))
df = lists_to_data(Y, X)
lst = data_to_lists(df)

}
