% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{inputs}
\alias{inputs}
\title{Function to get inputs for fused lasso fit}
\usage{
inputs(
  Y,
  X,
  weighting = FALSE,
  weights = NULL,
  wtype = "size",
  buildD = TRUE,
  normalize = TRUE
)
}
\arguments{
\item{Y}{list of size K. Its kth entry is a vector of size n_k}

\item{X}{list of size K. Its kth entry is a matrix of size n_kxp}

\item{weighting}{logical. If TRUE: use weighted regularization parameter}

\item{weights}{provide vector of weights}

\item{wtype}{either "size" (default, weighting based on group samle sizes)}

\item{buildD}{logical. If TRUE: compute Dmatrix}

\item{normalize}{logical. Whether or not to normalize the weights (currently divides by maximum).}
}
\value{
A list with the following components
\item{Y}{Response vector}
\item{X}{Block diagonal design matrix}
\item{D}{Coupling matrix}
\item{var_indicator}{pxK : indicates corresponding column (in the D-matrix) of each variable in each class}
}
\description{
Function to get inputs for fused lasso fit
}
