% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_lists.R
\name{lists_to_data}
\alias{lists_to_data}
\title{Take lists and convert them to data frame}
\usage{
lists_to_data(Y, X)
}
\arguments{
\item{Y}{list of response vectors}

\item{X}{list of predictor matrices}
}
\value{
A data frame with all the X variables, 
  as well as the Y variable and a new column identifying the class.
}
\description{
Take lists and convert them to data frame
}
