% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass.R
\name{mc_reg}
\alias{mc_reg}
\title{Wrapper function for easy multiclass regression}
\usage{
mc_reg(
  df,
  class_var = "class",
  y_var = "y",
  scale = TRUE,
  center_y = TRUE,
  lambdas = NULL,
  ...
)
}
\arguments{
\item{df}{data frame}

\item{class_var}{the name of the class variable in the data frame}

\item{y_var}{the name of the dependent variable in the data frame}

\item{scale}{whether or not to scale the predictors (logical, default = TRUE)}

\item{center_y}{whether or not to center the response (logical, default = TRUE)}

\item{...}{additional arguments to be passed on to multiclass_reg()}
}
\value{
A list with the following components
\item{fit}{fitted \code{genlasso} object}
\item{coef}{matrix of estimated coefficients}
\item{lambda}{vector of regularization parameters along knots of regularization path}
\item{K}{number of classes}
\item{p}{number of predictors per class}
\item{n}{vector of sample sizes for each class}
\item{Y}{composite Y vector (stacked over the classes)}
\item{X}{block diagonal X matrix (each block coresponds to a class)}
\item{var_indicator}{matrix of variable indicators}
}
\description{
A wrapperfunction that works with data frames, created to not ruin 
existing code that works directly with the multiclass_reg() function.
}
\examples{
p = 7
k = 2
n = 20
beta = c(1,2)
set.seed(1)
X = list(matrix(rnorm(p*n), ncol = p), matrix(rnorm(p*n), ncol = p))
Y = list(rnorm(n), rnorm(n))
df = lists_to_data(Y, X)
Xk <- mvtnorm::rmvnorm(n*2, mean=1:p, sigma=diag(1:p)) \%>\% round(2)
yk <- Xk \%*\% c(0,0,4,0,0,8,0) + rnorm(n*2) + rep(c(0,10), each = n) \%>\% round(2)
df = data.frame(y = yk, Xk, class = rep(c("A", "B"), each = n))
# force a missing predictor
df$X1[df$class == "A"] = NA
df$X7[df$class == "B"] = NA
mc = mc_reg(df)
 
}
