% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{multiclass_cv}
\alias{multiclass_cv}
\title{Function to compute optimal value of regularization parameter based on K-fold cross-validation}
\usage{
multiclass_cv(
  fit,
  Kfold = 5,
  nweight = NULL,
  weighting = TRUE,
  wtype = "size",
  type = "MSFE",
  l1length = 100,
  normalize = TRUE
)
}
\arguments{
\item{fit}{fitted object returned from multiclass_reg()}

\item{Kfold}{K-fold cross-validation}

\item{nweight}{vector of length K indicating weights for MSFE measure}

\item{type}{MSFE or MAFE}

\item{l1length}{length sparsity grid l1 penalty}

\item{asym}{flag: asymmetric forecast error measure or not}
}
\value{
A list with the following components
\item{lambda_opt}{Selected value of the regularization parameter via BIC}
\item{bhat_opt}{Estimated coefficients selected via BIC}
}
\description{
Function to compute optimal value of regularization parameter based on K-fold cross-validation
}
\examples{
p <- 7
k <- 2
n <- 50
beta <- c(1, 2)
set.seed(1)
X <- list(matrix(rnorm(p * n), ncol = p), matrix(rnorm(p * n), ncol = p))
Y <- list(rnorm(n), rnorm(n))
res <- multiclass_reg(Y, X)
cv_res <- multiclass_cv(res)
}
