% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass.R
\name{multiclass_reg}
\alias{multiclass_reg}
\title{Fit the multiclass regression estimator}
\usage{
multiclass_reg(
  Y,
  X,
  weighting = FALSE,
  weights = NULL,
  wtype = "size",
  initestim = NULL,
  normalize = TRUE,
  x_names = NULL,
  lambdas = NULL
)
}
\arguments{
\item{Y}{list of size K. Its kth entry is a vector of size n_k}

\item{X}{list of size K. Its kth entry is a matrix of size n_kxp}

\item{weighting}{logical. If TRUE: use weighted regularization parameter}

\item{weights}{provide vector of weights}

\item{wtype}{either "size" (default, weighting based on group samle sizes), "init" (weighting based on initial estimator), "combi" (weighting based on group sample sizes and initial estimator)}

\item{initestim}{pK-dimensional vector with estimated coefficients of initial estimator}

\item{normalize}{logical. Whether or not to normalize the weights (currently divides by maximum).}
}
\value{
A list with the following components
\item{fit}{fitted \code{genlasso} object}
\item{coef}{matrix of estimated coefficients}
\item{lambda}{vector of regularization parameters along knots of regularization path}
\item{K}{number of classes}
\item{p}{number of predictors per class}
\item{n}{vector of sample sizes for each class}
\item{Y}{composite Y vector (stacked over the classes)}
\item{X}{block diagonal X matrix (each block coresponds to a class)}
\item{var_indicator}{matrix of variable indicators}
}
\description{
Fit the multiclass regression estimator
}
\examples{
p = 7
k = 2
n = 20
beta = c(1,2)
set.seed(1)
X = list(matrix(rnorm(p*n), ncol = p), matrix(rnorm(p*n), ncol = p))
Y = list(rnorm(n), rnorm(n))
# no missing data
res = multiclass_reg(Y, X)
}
