% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass.R
\name{plot.mcreg}
\alias{plot.mcreg}
\title{Plot coefficient paths for multiclass regression objects}
\usage{
\method{plot}{mcreg}(
  obj,
  class_var = "class",
  x_vars = NULL,
  coef_type = c("rescaled", "scaled"),
  plot_type = c("knot", "lambda"),
  log_x = NULL
)
}
\arguments{
\item{obj}{an multiclass regression object (object of class mcreg)}

\item{class_var}{the name of the class variable in the data frame}

\item{y_var}{the name of the dependent variable in the data frame}

\item{...}{additional arguments to be passed on to multiclass_reg()}
}
\value{
A list with the following components
\item{fit}{fitted \code{genlasso} object}
\item{coef}{matrix of estimated coefficients}
\item{lambda}{vector of regularization parameters along knots of regularization path}
\item{K}{number of classes}
\item{p}{number of predictors per class}
\item{n}{vector of sample sizes for each class}
\item{Y}{composite Y vector (stacked over the classes)}
\item{X}{block diagonal X matrix (each block coresponds to a class)}
\item{var_indicator}{matrix of variable indicators}
}
\description{
Plot the coefficient paths.
}
\examples{
p = 7
k = 2
n = 20
beta = c(1,2)
set.seed(1)
X = list(matrix(rnorm(p*n), ncol = p), matrix(rnorm(p*n), ncol = p))
Y = list(rnorm(n), rnorm(n))
df = lists_to_data(Y, X)
mc_reg(df)
}
