% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass.R
\name{predict.mcreg}
\alias{predict.mcreg}
\title{Predict method for mcreg objects}
\usage{
\method{predict}{mcreg}(
  obj,
  class_var = "class",
  x_vars = NULL,
  coef_type = c("rescaled", "scaled"),
  plot_type = c("knot", "lambda"),
  log_x = NULL
)
}
\arguments{
\item{obj}{an multiclass regression object (object of class mcreg)}

\item{class_var}{the name of the class variable in the data frame}

\item{y_var}{the name of the dependent variable in the data frame}

\item{...}{additional arguments to be passed on to multiclass_reg()}
}
\value{
A list with the following components
}
\description{
Predict given an input.
}
\examples{
p = 7
k = 2
n = 20
beta = c(1,2)
set.seed(1)
X = list(matrix(rnorm(p*n), ncol = p), matrix(rnorm(p*n), ncol = p))
Y = list(rnorm(n), rnorm(n))
df = lists_to_data(Y, X)
multiclass_reg(df)
}
