% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{multiclass_cv}
\alias{multiclass_cv}
\title{Depricated function to compute optimal value of regularization parameter based on K-fold cross-validation}
\usage{
multiclass_cv(
  fit,
  Kfold = 5,
  nweight = NULL,
  weighting = TRUE,
  wtype = "size",
  type = "MSFE",
  l1length = 100,
  normalize = TRUE
)
}
\arguments{
\item{fit}{fitted object returned from multiclass_reg()}

\item{Kfold}{K-fold cross-validation}

\item{nweight}{vector of length K indicating weights for MSFE measure}

\item{type}{MSFE or MAFE}

\item{l1length}{length sparsity grid l1 penalty}

\item{asym}{flag: asymmetric forecast error measure or not}
}
\value{
A list with the following components
\item{lambda_opt}{Selected value of the regularization parameter via BIC}
\item{bhat_opt}{Estimated coefficients selected via BIC}
}
\description{
Depricated function to compute optimal value of regularization parameter based on K-fold cross-validation
}
